requirements.txt
Cython==3.0.9
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pytest==8.1.1
screeninfo==0.8.1
tomli==2.0.1
